requirements.txt
attrs==21.2.0
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.0.1
iniconfig==1.1.1
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.17.1
SQLAlchemy==1.4.15
toml==0.10.2
typing-extensions==3.10.0.0
Werkzeug==1.0.1
XlsxWriter==1.4.2
zipp==3.4.1
